fix(register): guard /register against already-authenticated visitors
/register had no guard at all, unlike /login's loginGuard — anyone already
signed in (any role, any registrationStatus) could still reach the
account-creation form. Extract the shared logic into guestOnlyGuard and add
registerGuard for /register.

Guarding via a /register/layout.tsx would have also caught /register/club/**,
which guards itself per-step and must stay reachable by an authenticated CLUB
user mid-registration — so instead split page.tsx into a thin async server
component (calls registerGuard) rendering a new client _components/RegisterForm.tsx
(the actual form, unchanged apart from the move).
